feat: add feed

+++ src/main/java/ru/yandex/practicum/storage/db/FilmDbStorage.java
@@ -7,6 +7,7 @@ import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.jdbc.support.rowset.SqlRowSet;
 import org.springframework.stereotype.Component;
+import ru.yandex.practicum.exception.AlreadyExistException;
 import ru.yandex.practicum.exception.UnknownDataException;
 import ru.yandex.practicum.model.film.Director;
 import ru.yandex.practicum.model.film.Film;
@@ -208,11 +209,19 @@ public class FilmDbStorage implements FilmStorage {
 
     @Override
     public void addLike(int filmId, int userId) {
+        if (isLikeExists(filmId, userId)){
+            log.warn("У фильма с идентификатором {} уже стоит лайк от пользователя с идентификатором {}.", filmId, userId);
+            throw new AlreadyExistException("У фильма уже стоит лайк от данного пользователя.");
+        }
         jdbcTemplate.update("INSERT INTO FILM_LIKES VALUES (?, ?)", filmId, userId);
     }
 
     @Override
     public void deleteLike(int filmId, int userId) {
+        if (!isLikeExists(filmId, userId)){
+            log.warn("У фильма с идентификатором {} нет лайка от пользователя с идентификатором {}.", filmId, userId);
+            throw new AlreadyExistException("Невозможно удалит лайк от пользователя от которого нет лайка");
+        }
         jdbcTemplate.update("DELETE FROM FILM_LIKES WHERE FILM_ID = ? AND USER_ID = ?", filmId, userId);
     }
 
@@ -355,5 +364,12 @@ public class FilmDbStorage implements FilmStorage {
             log.warn("Фильм с идентификатором {} не найден.", id);
             throw new UnknownDataException("Фильм c id = " + id + " не найден");
         }
+    }
+
+    @Override
+    public boolean isLikeExists ( int filmId, int userId){
+        String checkQuery = "SELECT * FROM FILM_LIKES WHERE FILM_ID = ? AND USER_ID = ? ";
+        SqlRowSet filmRows = jdbcTemplate.queryForRowSet(checkQuery, filmId, userId);
+        return filmRows.next();
     }
 }
